// ColtUI/src/app/shared/model/userCreate.ts
/**
 * FastAPI
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface UserCreate { 
    email: string;
    first_name: string;
    last_name: string;
    birthday: string;
    street: string;
    house_number: number;
    postcode: number;
    city: string;
}